pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "python-dotenv>=1.0.1",
    "solana>=0.36.5",
    "web3>=7.8.0",
]
